backend/requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
psycopg2
alembic
uvicorn[standard]